#  Optimize fetch subnet slaves recursive