--- a/functions/classes/class.Subnets.php
+++ b/functions/classes/class.Subnets.php
@@ -876,32 +876,66 @@ class Subnets extends Common_functions {
 	 * @param mixed $subnetId
 	 * @return void
 	 */
-	public function fetch_subnet_slaves_recursive ($subnetId) {
-		$end = false;							//loop break flag
-		# slaves array of id's, add current
-		$this->slaves[] = (int) $subnetId;		//id
+	public function fetch_subnet_slaves_recursive (int $subnetId) {
 
-		# loop
-		while($end == false) {
-			# fetch all immediate slaves
-			$slaves2 = $this->fetch_subnet_slaves ($subnetId);
+		// MySQL does not support multiple references to a temporary table in the same query.
+		// Ideally we would populate temp_table with our initial subnetId and then run the query below until
+		// the inserted record count is zero (no more slaves ids found).
+		//
+		//  INSERT INTO temp_table SELECT id FROM subnets WHERE id NOT IN (temp_table) AND masterSubnetId IN (temp_table)
+		//
+		// Work around the 'wont-fix' limitation by using a temporary table per iteration and consolidate the results.
 
-			# we have more slaves
-			if($slaves2) {
-				# recursive
-				foreach($slaves2 as $slave) {
-					# save to full array of slaves
-					$this->slaves_full[$slave->id] = $slave;
-					$this->slaves[] = $slave->id;
-					# fetch possible new slaves
-					$this->fetch_subnet_slaves_recursive ($slave->id);
-					$end = true;
-				}
+		$rowCount = 0; $level = 0;
+
+		try {
+			$this->Database->runQuery("DROP TABLE IF EXISTS tmp_subnet_familytree_${subnetId}_${level};");
+
+			$query = "CREATE TEMPORARY TABLE tmp_subnet_familytree_${subnetId}_${level} (id int(11) PRIMARY KEY, level int) ENGINE = MEMORY
+				SELECT id, 0 FROM subnets WHERE masterSubnetId = $subnetId;";
+			$result = $this->Database->runQuery($query, null, $rowCount);
+
+			// Fetch next level of slaves.
+			while ($result == 1 && $rowCount > 0){
+				$lastlevel = $level++;
+
+				$this->Database->runQuery("DROP TABLE IF EXISTS tmp_subnet_familytree_${subnetId}_${level};");
+
+				$query = "CREATE TEMPORARY TABLE tmp_subnet_familytree_${subnetId}_${level} (id int(11) PRIMARY KEY, level int) ENGINE = MEMORY
+					SELECT id, ${level} FROM subnets WHERE masterSubnetId IN (SELECT id FROM tmp_subnet_familytree_${subnetId}_${lastlevel});";
+				$result = $this->Database->runQuery($query, null, $rowCount);
 			}
-			# no more slaves
-			else {
-				$end = true;
+
+			// Consolidate results into the level 0 tempoary table.
+			while (--$level > 0) {
+				$query = "INSERT IGNORE INTO tmp_subnet_familytree_${subnetId}_0
+					SELECT * FROM tmp_subnet_familytree_${subnetId}_${level};";
+				$this->Database->runQuery($query);
+
+				$this->Database->runQuery("DROP TABLE IF EXISTS tmp_subnet_familytree_${subnetId}_${level};");
 			}
+
+			// Fetch all slaves
+			$slaves = $this->Database->getObjectsQuery("SELECT * FROM subnets WHERE id IN (SELECT id FROM tmp_subnet_familytree_${subnetId}_0);");
+
+			$this->Database->runQuery("DROP TABLE IF EXISTS tmp_subnet_familytree_${subnetId}_0;");
+
+			$this->slaves[] = $subnetId;
+
+			if (!$slaves) { return; }
+
+			foreach($slaves as $slave) {
+				# save to subnets cache
+				unset($slave->subnet_int);
+				$this->cache_write("subnets", $slave->id, $slave);
+
+				# save to full array of slaves
+				$this->slaves_full[$slave->id] = $slave;
+				$this->slaves[] = (int) $slave->id;
+			}
+		}
+		catch (Exception $e) {
+			$this->Result->show("danger", _("Error: ").$e->getMessage());
 		}
 	}