@GaryAllan 2017-05-06
The approach is to build a temporary table containing a list of all slave Ids for a given subnet and then use that temporary table to build SQL queries to fetch the information we need. This bulk request method significantly reduces the total number of SQL queries to the database and the total amount of data retrieved.

No changes to the database schema are required.

Temporary tables are locally scoped to the current MySQL connection and are automatically deleted when the MySQL connection is terminated (Usually at end of each .php file). Two or more .php files can safely run concurrently using the same temporary table names.

I'm using the MEMORY engine for added performance. Tables sizes using the MEMORY engine are restricted by the "max_heap_table_size" my.cnf variable. Storing 20,000 slave id values consumes approximately 1Mb and the default value for max_heap_table_size is 16M (MySQL 5.1). Scaling above 320,000 slave subnets would require increasing the "max_heap_table_size" my.cnf variable.

I've tested that the optimised and original fetch_subnet_slaves_recursive() function return equivalent results across all subnetId's in our database and a generated database containing 8000 IPv4 and IPv6 nested subnets.

I have pending submissions for optimisations on a few other Subnet and Address functions with similar performance improvements, these require Subnets::create_subnet_familytree() to be available so I'm waiting for review and feedback on this submission. (I'm happy to rename this function.)

The PHP profile results below were generated on an Intel(R) Core(TM)2 Duo CPU T5670 @ 1.80GHz displaying the subnet details of 10.0.0.0/8 (~5,500 nested subnets.)

http://127.0.0.1/?page=subnets&section=1&subnetId=1006

Original:  148s Total, 128s consumed by Subnets::fetch_subnet_slaves_recursive()
Patched:  19s Total, 4.2s consumed by Subnets::fetch_subnet_slaves_recursive()

![screenshot from 2017-05-06 14-24-03](https://cloud.githubusercontent.com/assets/18753294/25772743/5c661c9c-3269-11e7-81e4-44e9725d17a0.png)

![screenshot from 2017-05-06 14-33-46](https://cloud.githubusercontent.com/assets/18753294/25772746/63e6a504-3269-11e7-8b3f-4ef567a327c8.png)


I've got a patch set in development that caches the results of Subnets::create_subnet_familytree() and Subnets::calculate_subnet_usage into InnoDB tables. This gives a further 10x speed increase and would be useful where subnet/address changes are fairly infrequent. This would require schema changes and I'll submit patches for review at a later date after I've developed them further.

Thanks