@GaryAllan 2017-05-07
I should add the results above are with the default my.cnf value query_cache_type=0. Enabling the MySQL query cache does improve the timings significantly. (Your mileage may vary dependant upon your db structure, MySQL version and hardware.)

The timings below are with the query cache enabled (MariaBD 10.1.21) on the same hardware loading the same subnet details page. Cold cache was measured after running RESET QUERY CACHE to empty the cache. Warm cache results are timings from subsequent runs. 

    [mysqld]
    query_cache_type=1
    query_cache_size=64M
    query_cache_limit=8M

Original: 
Cache disabled: 148s Total, 128s consumed by Subnets::fetch_subnet_slaves_recursive()
Cold cache: 54s Total, 37.6s consumed by Subnets::fetch_subnet_slaves_recursive()
Warm cache:12s Total, 6.4s consumed by Subnets::fetch_subnet_slaves_recursive()

Patched: 
Cache disabled: 19s Total, 4.2s consumed by Subnets::fetch_subnet_slaves_recursive()
Cold cache: 15s Total, 4.2s consumed by Subnets::fetch_subnet_slaves_recursive()
Warm cache: 10s Total, 4.2s consumed by Subnets::fetch_subnet_slaves_recursive()

We're currently generating a very large number of queries to render webpages. [~37,000 to render the web page http://127.0.0.1/?page=subnets&section=1&subnetId=1006].

    MariaDB [phpipam]> show status like 'qcache%';
    +-------------------------+----------+
    | Variable_name           | Value    |
    +-------------------------+----------+
    | Qcache_free_blocks      | 1        |
    | Qcache_free_memory      | 46153888 |
    | Qcache_hits             | 170395   |
    | Qcache_inserts          | 26824    |
    | Qcache_lowmem_prunes    | 0        |
    | Qcache_not_cached       | 346      |
    | Qcache_queries_in_cache | 10697    |
    | Qcache_total_blocks     | 21410    |
    +-------------------------+----------+
    8 rows in set (0.01 sec)