@phpipam 2017-06-07
Hi, I added `set_tmptable_engine_type` that sets engine type from config.php, defaults to MEMORY, as a failsafe if someone does not have MEMORY storage.

Tested on 2 instances and it seems to be ok.